cleanup: fixes for makeallplat to pass with msp430-elf
Not all examples build: TI changed the header so CC430F5137 cannot
detect its broken CRC implementation; msp430-elf creates bigger images
so some applications overflow flash; some examples still depend on
mspgcc features like <sys/crtld.h>.

+++ examples/rf/cc3000/cli/main.c
@@ -1598,7 +1598,7 @@ cmd_ghbn (const char * argstr)
 {
   size_t argstr_len = strlen(argstr);
   const char * name;
-  uint16_t name_len;
+  size_t name_len;
   uint32_t addr = -1;
   int rc;
